backend/requirements.txt
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43
pandas>=2.1
numpy>=1.24
fastapi>=0.104
uvicorn[standard]>=0.24
joblib>=1.3
pydantic>=2.5
python-multipart>=0.0.6
imbalanced-learn>=0.12
pytorch-tabnet>=4.1
torch>=2.2
optuna>=3.6
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.2
